multiply

+++ Trapezium.py
@@ -1,6 +1,6 @@
 #the Area of Trapezium 1/2(a+b)h
-a = int(input('Enter Value of a'))
-b = int(input('Enter Value of b'))
-h = int(input('Enter Value of h(Height)'))
+a = float(input('Enter Value of a'))
+b = float(input('Enter Value of b'))
+h = float(input('Enter Value of h(Height)'))
 area = 1/2 * (a+b) * h
 print('Area of Trapezium is' , area)
